+++ NEWS.md
@@ -1,6 +1,8 @@
 # lava 1.9.2
   Development version
 
+  - new `c.summary.estimate` S3-method for concatenating `summary.estimate`
+    objects
   - sim.default can now operate on function return objects
     of type `estimate.extra` which are constructed via
     `c.estimate(estimate_object, extra_args)`